Fix Architecture
