+++ src/main/java/com/udemy/videolist/application/response/VideoUpdateResponse.java
@@ -1,20 +1,18 @@
-package com.udemy.videolist.application.controller;
+package com.udemy.videolist.application.response;
 
 import com.udemy.videolist.domain.model.Video;
 
-record VideoCreateResponse(
+public record VideoUpdateResponse(
     String message,
-    int id,
     String title,
     String instructor,
     String language,
     Boolean isFree,
     int price) {
 
-  public VideoCreateResponse(Video video) {
+  public VideoUpdateResponse(Video video) {
     this(
-        "video successfully created",
-        video.getId(),
+        "video successfully updated",
         video.getTitle(),
         video.getInstructor(),
         video.getLanguage(),
